add stubs for FontInterface, ImageInterface, EventInterface, GraphicsInterface and SettingsInterface postpone moving SDL code into the GraphicsInterface and move on with restructuring forbid creation of classes like Game, Scene etc. by hand. instead create Builder classes which does that start to implement Utils classes for all these classes mentioned above to separate different functionalities
TODO: add remaining Utils classes. get the entity element to work again. write unittests for builders and utils

+++ include/bkengine/utils/Timer.h
@@ -1,15 +1,16 @@
 #ifndef BKENGINE_TIMER_H
 #define BKENGINE_TIMER_H
 
-#include "SDLWrapper.h"
+#include <chrono>
+
 
 namespace bkengine
 {
     class Timer
     {
         private:
-            int startTicks;
-            int pausedTicks;
+            uint64_t startTicks;
+            uint64_t pausedTicks;
             bool paused;
             bool started;
 
@@ -19,7 +20,7 @@ namespace bkengine
             void stop();
             void pause();
             void unpause();
-            int getTicks() const;
+            uint64_t getTicks() const;
             bool isStarted() const;
             bool isPaused() const;
     };
